move vocab form to component and add group active reducer
