+++ src/redux/Actions/form.js
@@ -1,4 +1,8 @@
-import { SET_VOCAB_ACTIVE, SET_VOCAB_ACTIVATE } from "./index.js";
+import {
+  SET_VOCAB_ACTIVE,
+  SET_VOCAB_ACTIVATE,
+  SET_GROUP_ACTIVATE,
+} from "./index.js";
 
 export const setVocabActive = ({ active }) => {
   return {
@@ -14,6 +18,15 @@ export const setVocabActivate = ({ activate }) => {
     type: SET_VOCAB_ACTIVATE,
     payload: {
       activate,
+    },
+  };
+};
+
+export const setGroupActive = ({ active }) => {
+  return {
+    type: SET_GROUP_ACTIVATE,
+    payload: {
+      active,
     },
   };
 };
