+++ src/redux/Actions/index.js
@@ -8,3 +8,4 @@ export const SET_MENU_STYLE = "SET_MENU_STYLE";
 
 export const SET_VOCAB_ACTIVE = "SET_VOCAB_ACTIVE";
 export const SET_VOCAB_ACTIVATE = "SET_VOCAB_ACTIVATE";
+export const SET_GROUP_ACTIVATE = "SET_GROUP_ACTIVATE";
